requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
discord.py==1.7.3
emoji==1.6.1
gTTS==2.2.3
idna==3.3
multidict==5.2.0
playsound==1.3.0
pycparser==2.21
pydub==0.25.1
PyNaCl==1.4.0
python-voicetext==0.3.1
requests==2.26.0
six==1.16.0
typing-extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
